Refactor: centraliza topBar e bottomBar em funções reutilizáveis e ajusta navegação
