+++ app/src/main/java/com/example/burnbook/PaginaPrincipal.kt
@@ -9,6 +9,7 @@ import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.offset
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.safeContentPadding
 import androidx.compose.foundation.layout.safeDrawingPadding
@@ -62,9 +63,9 @@ fun PaginaPrincipal (navController: NavController) {
             bottomBar(isDarkMode, navController)
         },
 
-    ) {
+        ) {
         // Box principal
-        innerPadding ->
+            innerPadding ->
         Box(
             modifier = Modifier
                 .fillMaxSize()
@@ -87,31 +88,26 @@ fun PaginaPrincipal (navController: NavController) {
 
 
 @Composable
-fun topBar(isDarkMode: Boolean, onToggle: () -> Unit) {
-
+fun topBar(isDarkMode: Boolean, onToggle: () -> Unit){
     val iconeTopo = if (isDarkMode) R.drawable.sol else R.drawable.lua_
-
 
     Surface (
         color = if (isDarkMode) Color(0xFFDE425C) else Color(0xFFF65B75),
         modifier = Modifier
             .fillMaxWidth()
-            .height(80.dp)
+            .height(90.dp)
+
 
     ) {
         Row (
             horizontalArrangement = Arrangement.SpaceBetween,
             verticalAlignment = Alignment.CenterVertically,
-            modifier = Modifier.padding(start = 15.dp, end = 15.dp)
+            modifier = Modifier.padding(start = 15.dp)
         ) {
-//            Icon(painter = painterResource(id = R.drawable.icone_burnbook),
-//                contentDescription = "Ícone do Burn Book",
-//                tint = Color.Unspecified,
-//                modifier = Modifier.height(40.dp)
-//            )
 
             Column (
-                verticalArrangement = Arrangement.spacedBy((-18).dp)
+                verticalArrangement = Arrangement.spacedBy((-18).dp),
+                modifier = Modifier.padding(top = 15.dp)
             ) {
 
                 Text(
@@ -131,12 +127,16 @@ fun topBar(isDarkMode: Boolean, onToggle: () -> Unit) {
             }
             IconButton(
                 onClick = onToggle,
-                modifier = Modifier.size(50.dp)
+                modifier = Modifier
+                    .size(65.dp)
+                    .padding(top = 15.dp)
             ) {
                 Image(
                     painter = painterResource(id = iconeTopo),
                     contentDescription = "Trocar modo",
-                    modifier = Modifier.size(35.dp)
+                    modifier = Modifier
+                        .size(40.dp)
+                        .offset(y = 7.dp)
                 )
             }
 
@@ -146,7 +146,7 @@ fun topBar(isDarkMode: Boolean, onToggle: () -> Unit) {
 }
 
 @Composable
-fun bottomBarFun(isDarkMode: Boolean, navController: NavController) {
+fun bottomBar(isDarkMode: Boolean, navController: NavController) {
 
     Surface (
         color = if (isDarkMode) Color(0xFFDE425C) else Color(0xFFF65B75),
@@ -197,12 +197,19 @@ fun bottomBarFun(isDarkMode: Boolean, navController: NavController) {
                 tint = Color.Unspecified,
                 modifier = Modifier.height(30.dp)
             )
-            Icon(
-                painter = painterResource(id = R.drawable.image_usuario),
-                contentDescription = "Ícone do usuário",
-                tint = Color.Unspecified,
-                modifier = Modifier.height(30.dp)
-            )
+            IconButton(
+                onClick = {
+                    navController.navigate("usuario")
+                },
+            ) {
+                Icon(
+                    painter = painterResource(id = R.drawable.image_usuario),
+                    contentDescription = "Ícone do usuário",
+                    tint = Color.Unspecified,
+                    modifier = Modifier.height(30.dp)
+                )
+            }
+
         }
     }
 }
@@ -295,9 +302,9 @@ fun cardPost (isDarkMode: Boolean) {
             ) {
                 Text(
                     text = "Lorem Ipsum is simply dummy text of the printing and typesetting " +
-                        "industry. Lorem Ipsum has been the industry's standard dummy text ever " +
-                        "since the 1500s, when an unknown printer took a galley of type and " +
-                        "scrambled it to make a type specimen book. Lorem Ipsum is simply dummy text" +
+                            "industry. Lorem Ipsum has been the industry's standard dummy text ever " +
+                            "since the 1500s, when an unknown printer took a galley of type and " +
+                            "scrambled it to make a type specimen book. Lorem Ipsum is simply dummy text" +
                             " of the printing and typesetting industry.",
                     modifier = Modifier.padding(vertical = 8.dp, horizontal = 8.dp),
                     color = if (isDarkMode) Color.White else Color(0xFF2C2C2C),
